confirm version 0.9
almost finish the first version

+++ Server/GDouDiZhuServer/src/DDZRoomManager.cpp
@@ -10,7 +10,7 @@ IGameRoom* DDZRoomManager::createRoom(uint8_t nGameType)
 	return nullptr;
 }
 
-uint8_t DDZRoomManager::getDiamondNeed(uint8_t nGameType, uint8_t nLevel, ePayRoomCardType payType )
+uint32_t DDZRoomManager::getDiamondNeed(uint8_t nGameType, uint8_t nLevel, ePayRoomCardType payType )
 {
 #ifdef _DEBUG
 	return 0;
