+++ Server/GMaJiangServer/src/MJRoomManager.h
@@ -5,5 +5,5 @@ class MJRoomManager
 {
 public:
 	IGameRoom* createRoom(uint8_t nGameType)override;
-	uint8_t getDiamondNeed(uint8_t nGameType, uint8_t nLevel, ePayRoomCardType payType )override;
+	uint32_t getDiamondNeed(uint8_t nGameType, uint8_t nLevel, ePayRoomCardType payType )override;
 };
